change some i32 to c_uint

+++ audionimbus/src/context.rs
@@ -1,5 +1,6 @@
 use crate::error::{to_option_error, SteamAudioError};
 use crate::version::SteamAudioVersion;
+use std::os::raw::c_uint;
 
 /// A context object, which controls low-level operations of Steam Audio.
 ///
@@ -169,6 +170,6 @@ bitflags::bitflags! {
 
 impl From<ContextFlags> for audionimbus_sys::IPLContextFlags {
     fn from(context_flags: ContextFlags) -> Self {
-        Self(context_flags.bits())
+        Self(context_flags.bits() as c_uint)
     }
 }
