+++ audionimbus/src/effect/reflection.rs
@@ -12,6 +12,7 @@ use crate::ffi_wrapper::FFIWrapper;
 use crate::geometry::{Scene, SceneParams};
 use crate::probe::ProbeBatch;
 use crate::simulation::BakedDataIdentifier;
+use std::os::raw::c_uint;
 
 #[cfg(doc)]
 use crate::simulation::BakedDataVariation;
@@ -578,7 +579,7 @@ bitflags::bitflags! {
 
 impl From<ReflectionsBakeFlags> for audionimbus_sys::IPLReflectionsBakeFlags {
     fn from(reflections_bake_flags: ReflectionsBakeFlags) -> Self {
-        Self(reflections_bake_flags.bits())
+        Self(reflections_bake_flags.bits() as c_uint)
     }
 }
